+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## Unreleased
  * SVG exporter improvements
+ * SVG importer improvements
  * Implemented Alpha/Luma mattes on SVG export
  * Find/replace in lottie gui code editor
  * Properly parse transforms with xplit x/y
+ * Krita importer
 
 ## 0.6.2 - 2020-04-30
  * Support for all importers / exporters in lottie GUI